+++ src/replication-protocol/downstream.ts
@@ -117,7 +117,10 @@ export async function startReplicationDownstream<RxDocType, CheckpointType = any
                 }
             }).then(() => {
                 state.events.active.down.next(false);
-                if (!state.firstSyncDone.down.getValue()) {
+                if (
+                    !state.firstSyncDone.down.getValue() &&
+                    !state.events.canceled.getValue()
+                    ) {
                     state.firstSyncDone.down.next(true);
                 }
             });